requirements.txt
Django>=2.1
mysqlclient>=1.3.13
wheel>=0.31.1
gunicorn>=19.9.0
psycopg2>=2.7.5
networkx>=2.2
numpy>=1.15.3
matplotlib>=3.0.0
Pillow>=5.3.0
requests
